+++ internal/provider/model_resource_meraki_appliance_ssids.go
@@ -235,7 +235,7 @@ func (data *ResourceApplianceSSIDs) fromBody(ctx context.Context, res meraki.Res
 // easily change across versions of the backend API.) For List/Set/Map attributes, the func only updates the
 // "managed" elements, instead of all elements.
 func (data *ResourceApplianceSSIDs) fromBodyPartial(ctx context.Context, res meraki.Res) {
-	for i := 0; i < len(data.Items); i++ {
+	for i := range data.Items {
 		parent := &data
 		data := (*parent).Items[i]
 		parentRes := &res