+++ internal/provider/model_resource_meraki_organization_licenses.go
@@ -101,7 +101,7 @@ func (data *ResourceOrganizationLicenses) fromBody(ctx context.Context, res mera
 // easily change across versions of the backend API.) For List/Set/Map attributes, the func only updates the
 // "managed" elements, instead of all elements.
 func (data *ResourceOrganizationLicenses) fromBodyPartial(ctx context.Context, res meraki.Res) {
-	for i := 0; i < len(data.Items); i++ {
+	for i := range data.Items {
 		parent := &data
 		data := (*parent).Items[i]
 		parentRes := &res