+++ internal/provider/model_resource_meraki_sensor_mqtt_brokers.go
@@ -102,7 +102,7 @@ func (data *ResourceSensorMQTTBrokers) fromBody(ctx context.Context, res meraki.
 // easily change across versions of the backend API.) For List/Set/Map attributes, the func only updates the
 // "managed" elements, instead of all elements.
 func (data *ResourceSensorMQTTBrokers) fromBodyPartial(ctx context.Context, res meraki.Res) {
-	for i := 0; i < len(data.Items); i++ {
+	for i := range data.Items {
 		parent := &data
 		data := (*parent).Items[i]
 		parentRes := &res