+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `meraki_organization_licenses` resource
 - Add `meraki_sensor_mqtt_brokers` resource
 - Add `meraki_wireless_ssids` resource
+- Add `meraki_organization_policy_objects` resource
 
 ## 1.3.0